+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Fixed
 
+- `sandstone/Scroller` to focus properly when item is selected and `editable` is given
 - `sandstone/VirtualList` to not lose focus when a focused item is removed by reduced `dataSize`
 
 ## [2.7.1] - 2023-06-02