+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `sandstone/Scroller` to focus properly when item is selected and `editable` is given
+- `sandstone/VirtualList` to not lose focus when a focused item is removed by reduced `dataSize`
+
 ## [2.7.1] - 2023-06-02
 
 ### Added