+++ CHANGELOG.md
@@ -7,10 +7,13 @@ The following is a curated list of changes in the Enact sandstone module, newest
 ### Added
 
 - `sandstone/Icon` supported icon list, adding new icons `exclamation`, `show`, and `hide`
+- `sandstone/QuickGuidePanels` prop `onClose`
 
 ### Fixed
 
+- `sandstone/QuickGuidePanels` to update a close button position properly in RTL locales
 - `sandstone/Scroller` to focus properly when item is selected and `editable` is given
+- `sandstone/VirtualList` to not lose focus when a focused item is removed by reduced `dataSize`
 
 ## [2.7.1] - 2023-06-02