+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 ### Added
 
 - `sandstone/Icon` supported icon list, adding new icons `exclamation`, `show`, and `hide`
+- `sandstone/Scroller` to support showing buttons when an item is focused and `editable` is given
 - `sandstone/QuickGuidePanels` prop `onClose`
 
 ### Fixed