+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [2.7.2] - 2023-06-30
+
+### Added
+
+- `sandstone/Icon` supported icon list, adding new icons `exclamation`, `show`, and `hide`
+- `sandstone/QuickGuidePanels` prop `onClose`
+
+### Fixed
+
+- `sandstone/QuickGuidePanels` to update a close button position properly in RTL locales
+- `sandstone/Scroller` to focus properly when item is selected and `editable` is given
+- `sandstone/VirtualList` to not lose focus when a focused item is removed by reduced `dataSize`
+
 ## [2.5.11] - 2023-06-07
 
 ### Fixed